% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_C.R
\name{beta_C_curve}
\alias{beta_C_curve}
\title{Calculate beta_C/ beta_Sn curve}
\usage{
beta_C_curve(x)
}
\arguments{
\item{x}{a site by species matrix}
}
\value{
a tibble
}
\description{
Calculate beta_C/ beta_Sn curve
}
\examples{
\donttest{
library(vegan)
data(BCI)
beta_dat<-beta_C_curve(BCI)
}

}
