% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_C.R
\name{beta_SN}
\alias{beta_SN}
\title{Calculate beta_Sn}
\usage{
beta_SN(x, N)
}
\arguments{
\item{x}{a site by species matrix}

\item{N}{an integer value}
}
\value{
a numeric value
}
\description{
Calculate beta_Sn
}
\examples{
\donttest{
library(vegan)
data(BCI)
beta_SN(BCI, 50)
}
}
