% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwise_beta.R
\name{beta_stand}
\alias{beta_stand}
\title{Calculate pairwise/ setwise beta diversities}
\usage{
beta_stand(x, func = list("beta_true", "beta_SN"), setsize = 2,
  summarise = T, resamples = 1000, max_combn = 1000, verbose = T,
  args)
}
\arguments{
\item{x}{A site by species matrix.}

\item{setsize}{Number of samples per subset.}

\item{summarise}{Return mean and variance of all betas?}

\item{resamples}{Number of samples used if possible combinations exceeds max_combn}

\item{max_combn}{Number of combinations allowed before resampling is used instead.}

\item{verbose}{Print notifications?}
}
\value{
A data frame.
}
\description{
This function makes all pairwise/ setwise combinations of sites in the community matrix x and calculates pairwise/ setwise beta diversities,
respectively. "Set" refers to a user-defined number of samples that will be used calculate gamma diversities. This is essentially
the number of plots in a sample-based rarefaction. It's default is 2 (i.e. pairs). The function accepts the following indices: "S_PIE", "S_n", "S_cov".
The latter two require a value for the standardisation of number of individuals (N_stand) and coverage (S_cov).
If these arguments are not supplied, the function will standardise to the number of individuals of the smallest sample (if extrapolate = FALSE),
or twice this number (if extrapolate = TRUE). For index = "S_cov" the corresponding expected coverages will be used if C_stand= NULL.
}
\details{
If the number of possible sets/ pairwise comparisons exceeds the value of max_combn. A random subsample of all combinations will
be drawn. The number of random samples can be adjusted using the argument resamples. It is not recommended increase the
value of max_combn. The function choose() "n over k" can be used to manually compute the number of possible subsets where
n is the number of samples/rows in x and k is setsize. If summarise = T, mean and variance of all setwise
comparisons are returned. Otherwise, the function returns all individual beta values.

The default value for setsize is 2 which means that pairwise beta diversities are returned.
Pairwise beta_Sn values are independent of the gamma_scale diversity and sampling effort.
Therefore, they allow for meaningfull comparisons between groups with different numbers of
samples (see Marion et al, 2017).
}
