% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_C.R
\name{beta_true}
\alias{beta_true}
\title{Calculate Beta diversity}
\usage{
beta_true(x, transformation = F)
}
\arguments{
\item{x}{a site by species matrix}

\item{transformation}{apply Jost's transformation?}
}
\value{
numeric.
}
\description{
Calculate Beta diversity
}
\examples{
\donttest{
library(vegan)
data(BCI)
beta_true(BCI)
}
}
