% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_C.R
\name{C_target}
\alias{C_target}
\title{Calculate the recommended maximum coverage value for the computation of beta_C from a site by species matrix}
\usage{
C_target(x, factor = 2)
}
\arguments{
\item{x}{a site by specie matrix}
}
\value{
numeric value
}
\description{
This returns the coverage of x at the gamma scale that corresponds to the smalles observed sample size at the alpha scale.
}
\examples{
\donttest{
library(vegan)
data(BCI)

# What is the largest possible C that I can use to calculate beta_C for my site by species matrix?
C_target(BCI)
}

}
