% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_C.R
\name{calc_PIE}
\alias{calc_PIE}
\title{Calculate PIE and Effective number of species (S_PIE)}
\usage{
calc_PIE(x, ENS = FALSE)
}
\arguments{
\item{x}{site by species matrix}

\item{ENS}{return effective number of species?}
}
\value{

}
\description{
This function was copied from  the package \href{https://github.com/MoBiodiv/mobr/}{mobr}.
}
