% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_C.R
\name{coverage}
\alias{coverage}
\title{Calculate sample coverage}
\usage{
coverage(m)
}
\arguments{
\item{m}{site by species matrix}
}
\value{
a vector
}
\description{
Sample coverage is calculated according to Chao & Jost (2012).
}
\examples{
\donttest{
library(vegan)
data(BCI)

# at the alpha scale
coverage(BCI)

# at the gamma scale
coverage(colSums(BCI))
}

}
