% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_C.R
\name{D0.hat}
\alias{D0.hat}
\title{Rarefaction curve (inter- and extrapolation)}
\usage{
D0.hat(x, m)
}
\arguments{
\item{x}{integer vector of abundances}

\item{m}{sample size. can be a vector}
}
\value{
numerical (vector)
}
\description{
Returns the expected number of species for a samplesize m. Interpolation and extrapolation is possible.
This function was taken from the rPackage iNEXT (https://github.com/JohnsonHsieh/iNEXT).
}
