% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction_curves.R
\name{rarefy_long}
\alias{rarefy_long}
\title{Calculate IBR curves at alpha and gamma scales and return them in a long format}
\usage{
rarefy_long(x)
}
\arguments{
\item{x}{a site by species abundance matrix}
}
\value{
a data.frame
}
\description{
Please load tidyverse before using this.
}
