% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwise_beta.R
\name{beta_stand}
\alias{beta_stand}
\title{Calculate pairwise/ setwise beta diversities}
\usage{
beta_stand(
  x,
  setsize = 2,
  func = list("beta_true"),
  args = NULL,
  summarise = T,
  max_combn = 10000,
  resamples = 1000,
  verbose = T
)
}
\arguments{
\item{x}{A site by species matrix.}

\item{setsize}{Number of samples per subset.}

\item{func}{a list of function names to be used for the metric calculation}

\item{args}{a list of additional arguments to be passed on to the functions in func}

\item{summarise}{Return mean and variance of all betas?}

\item{max_combn}{Number of combinations allowed before resampling is used instead}

\item{resamples}{Number of samples used if possible combinations exceeds max_combn}

\item{verbose}{Print notifications?}
}
\value{
A data frame
}
\description{
This function makes pairwise/ setwise combinations of sites in the community matrix x and calculates pairwise/ setwise beta diversities,
respectively. "Set" refers to a user-defined number of samples that will be used. This is essentially
the number of plots in a sample-based rarefaction. It's default is 2 (i.e. pairs).
}
\details{
The calulation of the beta-diversity metric is done according to the function in "func". This should be a function that returns a
(dis-)similarity or beta diversity metric for a site-by species matrix like x. "func" is applied to every subset. Addidtional arguments will be passed on to "func".

If the number of possible sets/ pairwise comparisons exceeds the value of max_combn. A random subsample of all combinations will
be drawn. The number of random samples can be adjusted using the argument resamples. It is not recommended increase the
value of max_combn. The function choose() "n over k" can be used to manually compute the number of possible subsets where
n is the number of samples/rows in x and k is setsize. If summarise = T, mean and variance of all setwise
comparisons are returned. Otherwise, the function returns all individual beta values.
}
\examples{
\donttest{
library(vegan)
data(BCI)
beta_pairwise<-
beta_stand(BCI, func = list("beta_true", "beta_SN", "beta_C"), setsize=2,
args = list(N=150, C=0.5))
}

}
