% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwise_beta.R
\name{setwise_beta}
\alias{setwise_beta}
\title{Calculate pairwise/ setwise beta diversities}
\usage{
setwise_beta(x, setsize = 2, func, ...)
}
\arguments{
\item{x}{A site by species matrix.}

\item{setsize}{number of sites in the matrix}

\item{func}{function returning the desired beta-diversity metric}

\item{...}{additional argumets passed on to func}
}
\value{
A single numeric value for the beta diversity index.
}
\description{
This is an internal helper function for beta_stand(). Beta-diversity is calculated for the number of samples
in x and the the given index. Therefore, x must have exactly N_samples rows. This is a security feature to ensure that users don't
try to calculate another set size than given by the number of rows in x. Use beta_stand()!
}
\details{
If extrapolate = F, observed S will be returned if N_stand exceeds sample size.
}
